style: rename menu-button file
